Refactored RoomToRoomNavigation to Extract Classes/Functions
Following classes were extracted from RoomToRoomNavigation for bette rmodularity and reusability.

- FloorPlanService
- NavigationSteps
- NavigationStylesService
- NavigationValidationService
- PathCalculationService
