+++ ConcordiaMaps/components/IndoorNavigation/RoomToRoomNavigation.js
@@ -1,13 +1,27 @@
 import React, { useState, useEffect, useRef } from "react";
-import { View, Text, TouchableOpacity, ScrollView, Modal } from "react-native";
+import { View, Text, TouchableOpacity, ScrollView, Alert } from "react-native";
 import { WebView } from "react-native-webview";
 import { useRoute, useNavigation } from "@react-navigation/native";
-import { findShortestPath } from "./PathFinder";
 import FloorRegistry from "../../services/BuildingDataService";
 import Header from "../Header";
 import NavBar from "../NavBar";
 import styles from "../../styles/IndoorNavigation/RoomtoRoomNavigationStyles";
-import ExpandedFloorPlanModal from "./ExpandedFloorPlanModal";
+import ExpandedFloorPlanModal from "./ExpandedFloorPlan";
+import NavigationSteps from "./NavigationSteps";
+
+// Import extracted services
+import { generateFloorHtml } from "../../services/FloorPlanService";
+import { 
+  findBuildingTypeFromId,
+  findFloorForRoom,
+  validateRoomSelection 
+} from "../../services/NavigationValidationService";
+import { 
+  calculateNavigationPath,
+  handleSameFloorNavigation,
+  handleInterFloorNavigation,
+  findTransportMethod
+} from "../../services/PathCalculationService";
 
 const RoomToRoomNavigation = () => {
   const route = useRoute();
@@ -53,33 +67,10 @@ const RoomToRoomNavigation = () => {
   // Available buildings
   const buildings = FloorRegistry.getBuildings();
 
-  // Function to find buildingType from buildingId
-  const findBuildingTypeFromId = (id) => {
-    const buildingTypes = Object.keys(FloorRegistry.getAllBuildings());
-    return buildingTypes.find(
-      (key) => FloorRegistry.getBuilding(key).id === id
-    );
-  };
-
-  // Function to find which floor a room is on
-  const findFloorForRoom = (buildingType, roomId) => {
-    const building = FloorRegistry.getBuilding(buildingType);
-    if (!building) return null;
-    
-    // Check each floor for the room
-    for (const floorId in building.floors) {
-      const rooms = FloorRegistry.getRooms(buildingType, floorId);
-      if (rooms && roomId in rooms) {
-        return floorId;
-      }
-    }
-    return null;
-  };
-
   // Initialize from route params on component mount
   useEffect(() => {
     if (buildingId) {
-      const determinedBuildingType = findBuildingTypeFromId(buildingId);
+      const determinedBuildingType = findBuildingTypeFromId(buildingId, FloorRegistry);
       
       if (determinedBuildingType) {
         console.log(`Found building type ${determinedBuildingType} for ID ${buildingId}`);
@@ -92,13 +83,13 @@ const RoomToRoomNavigation = () => {
 
         // If no start floor provided but have start room, find the floor
         if (!foundStartFloor && startRoom) {
-          foundStartFloor = findFloorForRoom(determinedBuildingType, startRoom);
+          foundStartFloor = findFloorForRoom(determinedBuildingType, startRoom, FloorRegistry);
           console.log(`Determined start floor: ${foundStartFloor} for room ${startRoom}`);
         }
 
         // If no end floor provided but have end room, find the floor
         if (!foundEndFloor && endRoom) {
-          foundEndFloor = findFloorForRoom(determinedBuildingType, endRoom);
+          foundEndFloor = findFloorForRoom(determinedBuildingType, endRoom, FloorRegistry);
           console.log(`Determined end floor: ${foundEndFloor} for room ${endRoom}`);
         }
 
@@ -288,153 +279,7 @@ const RoomToRoomNavigation = () => {
     }
   };
 
-  // Break out validation into separate function
-  const validateRoomSelection = (
-    startFloorGraph,
-    endFloorGraph,
-    selectedStartRoom,
-    selectedEndRoom,
-  ) => {
-    if (!selectedStartRoom || !selectedEndRoom) {
-      return "Please select both start and end rooms";
-    }
-
-    if (!startFloorGraph[selectedStartRoom]) {
-      return `Start room ${selectedStartRoom} not found in navigation graph`;
-    }
-
-    if (!endFloorGraph[selectedEndRoom]) {
-      return `End room ${selectedEndRoom} not found in navigation graph`;
-    }
-
-    return null;
-  };
-
-  // Handle same floor navigation
-  const handleSameFloorNavigation = (
-    startFloorGraph,
-    selectedStartRoom,
-    selectedEndRoom,
-    startFloor,
-    buildingName,
-  ) => {
-    const directPath = findShortestPath(
-      startFloorGraph,
-      selectedStartRoom,
-      selectedEndRoom,
-    );
-
-    if (directPath.length < 2) {
-      throw new Error("Could not find a path between these rooms");
-    }
-
-    return {
-      startFloorPath: directPath,
-      endFloorPath: [],
-      navigationSteps: [
-        {
-          type: "start",
-          text: `Start at room ${selectedStartRoom} on floor ${startFloor} of ${buildingName}`,
-        },
-        ...directPath.slice(1, -1).map((node) => ({
-          type: "walk",
-          text: `Go to ${node}`,
-        })),
-        {
-          type: "end",
-          text: `Arrive at destination: ${selectedEndRoom}`,
-        }
-      ],
-    };
-  };
-
-  // Find available transportation method
-  const findTransportMethod = (startFloorGraph, endFloorGraph) => {
-    const startNodes = new Set(Object.keys(startFloorGraph));
-    const endNodes = new Set(Object.keys(endFloorGraph));
-
-    const transportMethods = ["escalator", "elevator", "stairs"];
-
-    for (const method of transportMethods) {
-      if (startNodes.has(method) && endNodes.has(method)) {
-        return method;
-      }
-    }
-
-    return null;
-  };
-
-  // Handle inter-floor navigation
-  const handleInterFloorNavigation = (
-    startFloorGraph,
-    endFloorGraph,
-    selectedStartRoom,
-    selectedEndRoom,
-    startFloor,
-    endFloor,
-    buildingName,
-  ) => {
-    const transportMethod = findTransportMethod(startFloorGraph, endFloorGraph);
-
-    if (!transportMethod) {
-      throw new Error(
-        `Cannot navigate between floors ${startFloor} and ${endFloor}`,
-      );
-    }
-
-    const startFloorTransportPath = findShortestPath(
-      startFloorGraph,
-      selectedStartRoom,
-      transportMethod,
-    );
-    const endFloorTransportPath = findShortestPath(
-      endFloorGraph,
-      transportMethod,
-      selectedEndRoom,
-    );
-
-    if (
-      startFloorTransportPath.length < 2 ||
-      endFloorTransportPath.length < 2
-    ) {
-      throw new Error("Could not find a complete path between these rooms");
-    }
-
-    return {
-      startFloorPath: startFloorTransportPath,
-      endFloorPath: endFloorTransportPath,
-      navigationSteps: [
-        {
-          type: "start",
-          text: `Start at room ${selectedStartRoom} on floor ${startFloor} of ${buildingName}`,
-        },
-        ...startFloorTransportPath.slice(1).map((node, index) => ({
-          type: "walk",
-          text:
-            index === startFloorTransportPath.length - 2
-              ? `Arrive at ${transportMethod} on floor ${startFloor}`
-              : `Go to ${node}`,
-        })),
-        {
-          type: transportMethod,
-          text: `Take ${transportMethod} to floor ${endFloor}`,
-        },
-        ...endFloorTransportPath.slice(1).map((node, index) => ({
-          type: "walk",
-          text:
-            index === 0
-              ? `Start from ${transportMethod} on floor ${endFloor}`
-              : `Go to ${node}`,
-        })),
-        {
-          type: "end",
-          text: `Arrive at destination: ${selectedEndRoom}`,
-        },
-      ],
-    };
-  };
-
-  // Main calculatePath function with reduced complexity
+  // Main calculatePath function (refactored)
   const calculatePath = () => {
     try {
       // Verify all required data is available
@@ -459,7 +304,7 @@ const RoomToRoomNavigation = () => {
         startFloorGraph,
         endFloorGraph,
         selectedStartRoom,
-        selectedEndRoom,
+        selectedEndRoom
       );
 
       if (validationError) {
@@ -469,13 +314,14 @@ const RoomToRoomNavigation = () => {
       }
 
       let result;
+      // Use the appropriate path calculation function
       if (startFloor === endFloor) {
         result = handleSameFloorNavigation(
           startFloorGraph,
           selectedStartRoom,
           selectedEndRoom,
           startFloor,
-          building.name,
+          building.name
         );
       } else {
         console.log("Finding transport method between floors");
@@ -489,7 +335,7 @@ const RoomToRoomNavigation = () => {
           selectedEndRoom,
           startFloor,
           endFloor,
-          building.name,
+          building.name
         );
       }
 
@@ -514,225 +360,10 @@ const RoomToRoomNavigation = () => {
     }
   };
 
-  // Generate HTML for floor visualization with path
-  const generateFloorHtml = (floorPlan, pathNodes = [], rooms = {}) => {
-    // Prepare path data by converting node names to coordinates
-    const pathCoordinates = pathNodes
-      .map((node) => (rooms[node] ? rooms[node] : null))
-      .filter((coord) => coord !== null);
-
-    // Serialize path data for safe injection into HTML
-    const pathDataJson = JSON.stringify(pathCoordinates);
-
-    return `
-      <!DOCTYPE html>
-      <html>
-        <head>
-          <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=2.0, user-scalable=yes">
-          <style>
-            body, html {
-              margin: 0;
-              padding: 0;
-              height: 100%;
-              width: 100%;
-              overflow: hidden;
-              touch-action: manipulation;
-            }
-
-            #svg-container {
-              width: 100%;
-              height: 100%;
-              display: flex;
-              justify-content: center;
-              align-items: center;
-              overflow: hidden;
-              position: relative;
-              background-color: #f5f5f5;
-            }
-
-            svg {
-              max-width: 100%;
-              max-height: 100%;
-              width: auto !important;
-              height: auto !important;
-              display: block;
-              visibility: hidden; /* Start hidden and show when ready */
-            }
-
-            .navigation-path { 
-              fill: none;
-              stroke: #912338; 
-              stroke-width: 4; 
-              stroke-linecap: round;
-              stroke-linejoin: round;
-              stroke-dasharray: 10,5;
-              animation: dash 1s linear infinite;
-            }
-            
-            @keyframes dash {
-              to {
-                stroke-dashoffset: -15;
-              }
-            }
-            
-            .room-highlight {
-              fill: #912338;
-              fill-opacity: 0.5;
-              stroke: #912338;
-              stroke-width: 2;
-              rx: 5;
-              ry: 5;
-            }
-            
-            /* Loading indicator */
-            #loader {
-              border: 5px solid #f3f3f3;
-              border-top: 5px solid #912338;
-              border-radius: 50%;
-              width: 30px;
-              height: 30px;
-              animation: spin 1s linear infinite;
-              position: absolute;
-              top: 50%;
-              left: 50%;
-              margin-top: -15px;
-              margin-left: -15px;
-              z-index: 100;
-            }
-            
-            @keyframes spin {
-              0% { transform: rotate(0deg); }
-              100% { transform: rotate(360deg); }
-            }
-          </style>
-          <script>
-            // Make initialization more robust with retry mechanism
-            let initAttempts = 0;
-            const maxAttempts = 5;
-            let svgInitialized = false;
-            
-            document.addEventListener('DOMContentLoaded', function() {
-              const loader = document.getElementById('loader');
-              
-              function initializeSVG() {
-                if (svgInitialized) return; // Prevent multiple initializations
-                
-                const svg = document.querySelector('svg');
-                if (!svg && initAttempts < maxAttempts) {
-                  console.log('SVG element not found, retry attempt ' + (initAttempts + 1));
-                  initAttempts++;
-                  setTimeout(initializeSVG, 200);
-                  return;
-                }
-                
-                if (!svg) {
-                  console.error('SVG element not found after ' + maxAttempts + ' attempts');
-                  if (loader && loader.parentNode) {
-                    loader.parentNode.removeChild(loader);
-                  }
-                  return;
-                }
-                
-                try {
-                  // Remove loading indicator
-                  if (loader && loader.parentNode) {
-                    loader.parentNode.removeChild(loader);
-                  }
-
-                  // Set viewBox if not already set
-                  if (!svg.getAttribute('viewBox')) {
-                    const bbox = svg.getBBox();
-                    svg.setAttribute('viewBox', \`\${bbox.x} \${bbox.y} \${bbox.width} \${bbox.height}\`);
-                  }
-                  
-                  // Set preserveAspectRatio to see the full floor plan
-                  svg.setAttribute('preserveAspectRatio', 'xMidYMid meet');
-                  
-                  // Function to visualize the path
-                  function visualizePath(coordinates) {
-                    if (!coordinates || coordinates.length < 2) {
-                      console.warn('Not enough coordinates to draw path');
-                      return;
-                    }
-                    
-                    // Create SVG path element
-                    const svgNS = "http://www.w3.org/2000/svg";
-                    const pathElement = document.createElementNS(svgNS, "path");
-                    pathElement.classList.add('navigation-path');
-                    
-                    // Build the path data string
-                    let pathData = '';
-                    
-                    coordinates.forEach((coord, index) => {
-                      if (!coord || !coord.nearestPoint) {
-                        console.warn('Invalid coordinate at index', index);
-                        return;
-                      }
-                      
-                      const point = coord.nearestPoint;
-                      if (index === 0) {
-                        // Move to the first point
-                        pathData += \`M \${point.x} \${point.y}\`;
-                      } else {
-                        // Line to subsequent points
-                        pathData += \`L \${point.x} \${point.y}\`;
-                      }
-                    });
-                    
-                    if (pathData === '') {
-                      console.warn('No valid path data could be generated');
-                      return;
-                    }
-                    
-                    // Set path attributes
-                    pathElement.setAttribute('d', pathData);
-                    
-                    // Add to SVG
-                    svg.appendChild(pathElement);
-                    
-                    console.log('Path drawn with', coordinates.length, 'points');
-                  }
-
-                  // Get path coordinates and draw
-                  const pathCoordinates = ${pathDataJson};
-                  visualizePath(pathCoordinates);
-                  
-                  // Make SVG visible after everything is ready
-                  svg.style.visibility = 'visible';
-                  svgInitialized = true;
-                } catch (error) {
-                  console.error('Error in SVG initialization:', error);
-                  if (svg) svg.style.visibility = 'visible'; // Show SVG even if there was an error
-                }
-              }
-              
-              // Initialize with a delay to ensure DOM is ready
-              setTimeout(initializeSVG, 300);
-            });
-            
-            // Prevent reinitializing on resize
-            window.addEventListener('resize', function() {
-              const svg = document.querySelector('svg');
-              if (svg) {
-                svg.setAttribute('preserveAspectRatio', 'xMidYMid meet');
-              }
-            });
-          </script>
-        </head>
-        <body>
-          <div id="svg-container">
-            <div id="loader"></div>
-            ${floorPlan || '<div style="color:red;padding:20px;text-align:center;">No SVG loaded</div>'}
-          </div>
-        </body>
-      </html>
-    `;
-  };
-
   // Render the building selection screen
   const renderBuildingSelection = () => (
     <View style={styles.sectionContainer}>
-      <Text style={styles.sectionTitle}>Select a Building</Text>
+      <Text style={styles.sectionTitle}>Select Building</Text>
       <ScrollView style={styles.scrollContainer}>
         {buildings.map((building) => (
           <TouchableOpacity
@@ -907,71 +538,34 @@ const RoomToRoomNavigation = () => {
     </View>
   );
 
-  // Render the navigation results screen
-const renderNavigation = () => {
-  const navigation = useNavigation();
-  const { skipSelection } = route.params || {};
+  // Render the navigation results screen (using NavigationSteps component)
+  const renderNavigation = () => {
+    const navigation = useNavigation();
+    const { skipSelection } = route.params || {};
 
-  return (
-    <View style={styles.sectionContainer}>
-      <Text style={styles.sectionTitle}>Navigation</Text>
+    return (
+      <View style={styles.sectionContainer}>
+        <Text style={styles.sectionTitle}>Navigation</Text>
 
-      <View style={styles.navigationContainer}>
-        <View style={styles.floorsContainer}>
-          <View style={styles.floorPlanContainer}>
-            <Text style={styles.floorColumnTitle}>Floor {startFloor}</Text>
-            <TouchableOpacity
-              style={styles.expandButton}
-              onPress={() => setExpandedFloor(startFloor)}
-            >
-              <Text style={styles.expandButtonText}>Expand</Text>
-            </TouchableOpacity>
-            <View style={styles.floorPlanWrapper}>
-              <WebView
-                ref={startFloorWebViewRef}
-                originWhitelist={["*"]}
-                source={{
-                  html: generateFloorHtml(
-                    startFloorPlan,
-                    startFloorPath,
-                    startFloorRooms,
-                  ),
-                }}
-                style={styles.floorPlan}
-                scrollEnabled={false}
-                onError={(e) =>
-                  console.error("WebView error:", e.nativeEvent)
-                }
-                onLoadEnd={() => console.log("WebView loaded")}
-                renderLoading={() => <View style={styles.webViewLoader}><Text>Loading...</Text></View>}
-                startInLoadingState={true}
-                javaScriptEnabled={true}
-                domStorageEnabled={true}
-                cacheEnabled={false}
-                incognito={true}
-                key={`start-${startFloor}-${startFloorPath.length}`} // Add a key to control remounting
-              />
-            </View>
-          </View>
-
-          {startFloor !== endFloor && endFloorPlan && (
+        <View style={styles.navigationContainer}>
+          <View style={styles.floorsContainer}>
             <View style={styles.floorPlanContainer}>
-              <Text style={styles.floorColumnTitle}>Floor {endFloor}</Text>
+              <Text style={styles.floorColumnTitle}>Floor {startFloor}</Text>
               <TouchableOpacity
                 style={styles.expandButton}
-                onPress={() => setExpandedFloor(endFloor)}
+                onPress={() => setExpandedFloor(startFloor)}
               >
                 <Text style={styles.expandButtonText}>Expand</Text>
               </TouchableOpacity>
               <View style={styles.floorPlanWrapper}>
                 <WebView
-                  ref={endFloorWebViewRef}
+                  ref={startFloorWebViewRef}
                   originWhitelist={["*"]}
                   source={{
                     html: generateFloorHtml(
-                      endFloorPlan,
-                      endFloorPath,
-                      endFloorRooms,
+                      startFloorPlan,
+                      startFloorPath,
+                      startFloorRooms,
                     ),
                   }}
                   style={styles.floorPlan}
@@ -986,89 +580,99 @@ const renderNavigation = () => {
                   domStorageEnabled={true}
                   cacheEnabled={false}
                   incognito={true}
-                  key={`end-${endFloor}-${endFloorPath.length}`} // Add a key to control remounting
+                  key={`start-${startFloor}-${startFloorPath.length}`} // Add a key to control remounting
                 />
               </View>
             </View>
-          )}
-        </View>
 
-        <View style={styles.stepsContainer}>
-          <Text style={styles.stepsTitle}>Navigation Steps</Text>
-          <ScrollView style={styles.stepsList}>
-            {navigationSteps.map((step, index) => (
-              <View key={index} style={styles.stepItem}>
-                <View
-                  style={[
-                    styles.stepDot,
-                    { backgroundColor: getStepColor(step.type) },
-                  ]}
-                />
-                <Text style={styles.stepText}>{step.text}</Text>
+            {startFloor !== endFloor && endFloorPlan && (
+              <View style={styles.floorPlanContainer}>
+                <Text style={styles.floorColumnTitle}>Floor {endFloor}</Text>
+                <TouchableOpacity
+                  style={styles.expandButton}
+                  onPress={() => setExpandedFloor(endFloor)}
+                >
+                  <Text style={styles.expandButtonText}>Expand</Text>
+                </TouchableOpacity>
+                <View style={styles.floorPlanWrapper}>
+                  <WebView
+                    ref={endFloorWebViewRef}
+                    originWhitelist={["*"]}
+                    source={{
+                      html: generateFloorHtml(
+                        endFloorPlan,
+                        endFloorPath,
+                        endFloorRooms,
+                      ),
+                    }}
+                    style={styles.floorPlan}
+                    scrollEnabled={false}
+                    onError={(e) =>
+                      console.error("WebView error:", e.nativeEvent)
+                    }
+                    onLoadEnd={() => console.log("WebView loaded")}
+                    renderLoading={() => <View style={styles.webViewLoader}><Text>Loading...</Text></View>}
+                    startInLoadingState={true}
+                    javaScriptEnabled={true}
+                    domStorageEnabled={true}
+                    cacheEnabled={false}
+                    incognito={true}
+                    key={`end-${endFloor}-${endFloorPath.length}`} // Add a key to control remounting
+                  />
+                </View>
               </View>
-            ))}
-          </ScrollView>
+            )}
+          </View>
+
+          {/* Use the extracted NavigationSteps component */}
+          <NavigationSteps steps={navigationSteps} />
         </View>
+
+        {/* Conditional rendering for back button based on how this screen was opened */}
+        {skipSelection ? (
+          <TouchableOpacity
+            style={styles.backButton}
+            onPress={() => navigation.goBack()}
+          >
+            <Text style={styles.backButtonText}>Return to Journey</Text>
+          </TouchableOpacity>
+        ) : (
+          <TouchableOpacity
+            style={styles.backButton}
+            onPress={() => setStep("rooms")}
+          >
+            <Text style={styles.backButtonText}>Back to Room Selection</Text>
+          </TouchableOpacity>
+        )}
       </View>
+    );
+  };
 
-      {/* Conditional rendering for back button based on how this screen was opened */}
-      {skipSelection ? (
-        <TouchableOpacity
-          style={styles.backButton}
-          onPress={() => navigation.goBack()}
-        >
-          <Text style={styles.backButtonText}>Return to Journey</Text>
-        </TouchableOpacity>
-      ) : (
-        <TouchableOpacity
-          style={styles.backButton}
-          onPress={() => setStep("rooms")}
-        >
-          <Text style={styles.backButtonText}>Back to Room Selection</Text>
-        </TouchableOpacity>
-      )}
-    </View>
-  );
-};
-
-
-const renderExpandedFloorPlan = () => {
-  if (!expandedFloor) return null;
-  
-  return (
-    <ExpandedFloorPlanModal
-      visible={!!expandedFloor}
-      floorNumber={expandedFloor}
-      onClose={() => setExpandedFloor(null)}
-      htmlContent={generateFloorHtml(floorPlan, pathNodes, rooms)}
-      webViewProps={{
-        scrollEnabled: false,
-        onError: (e) => console.error("WebView error in modal:", e.nativeEvent),
-        onLoadEnd: () => console.log("Modal WebView loaded"),
-        cacheEnabled: false,
-        incognito: true,
-        key: `expanded-${expandedFloor}-${pathNodes.length}`,
-      }}
-    />
-  );
-};
-
-  // Helper function to get color for navigation step type
-  const getStepColor = (type) => {
-    switch (type) {
-      case "start":
-        return "#4CAF50"; // Green
-      case "end":
-        return "#F44336"; // Red
-      case "escalator":
-        return "#2196F3"; // Blue
-      case "elevator":
-        return "#9C27B0"; // Purple
-      case "stairs":
-        return "#FF9800"; // Orange
-      default:
-        return "#912338"; // Maroon
-    }
+  // Render expanded floor plan modal
+  const renderExpandedFloorPlan = () => {
+    if (!expandedFloor) return null;
+    
+    const isStartFloor = expandedFloor === startFloor;
+    const floorPlan = isStartFloor ? startFloorPlan : endFloorPlan;
+    const pathNodes = isStartFloor ? startFloorPath : endFloorPath;
+    const rooms = isStartFloor ? startFloorRooms : endFloorRooms;
+    
+    return (
+      <ExpandedFloorPlanModal
+        visible={!!expandedFloor}
+        floorNumber={expandedFloor}
+        onClose={() => setExpandedFloor(null)}
+        htmlContent={generateFloorHtml(floorPlan, pathNodes, rooms)}
+        webViewProps={{
+          scrollEnabled: false,
+          onError: (e) => console.error("WebView error in modal:", e.nativeEvent),
+          onLoadEnd: () => console.log("Modal WebView loaded"),
+          cacheEnabled: false,
+          incognito: true,
+          key: `expanded-${expandedFloor}-${pathNodes.length}`,
+        }}
+      />
+    );
   };
 
   // Main render method
